Created an interface for ProductRepository and a class that implements ProductRepository and stores products in a list.
